move handle proposal funcs to app and clean up
